refactor: Introduce custom workspace manager
Replace FrameWebWorkspaceInitializer with the new FrameWebWorkspaceManager.

This new manager extends Langium's DefaultWorkspaceManager to provide more robust workspace handling. It is responsible for discovering and loading all framework-specific files (.fwt) recursively, ensuring the CLI correctly initializes the workspace and properly resolves the scope step. The default Workspace-manager is used on the extension.